создание класса логика
