+++ src/main/java/botLogic/Logic.java
@@ -1,0 +1,8 @@
+package botLogic;
+
+public class Logic{
+    public static String  processMessage(String message){
+        return message;
+    }
+}
+
