--- src/main/java/org/echoJavaTelegramBot/Main.java
@@ -1,15 +1,0 @@
-package org.echoJavaTelegramBot;
-
-import org.telegram.telegrambots.meta.TelegramBotsApi;
-import org.telegram.telegrambots.meta.api.objects.Update;
-import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
-import org.telegram.telegrambots.updatesreceivers.DefaultBotSession;
-
-public class Main {
-    public static void main(String[] args) {
-        String botName = "echo";
-        String botToken = System.getenv("BOT_TOKEN");
-        Bot tg_bot = new Telegram_Bot(botName, botToken);
-    }
-}
-
